test: consolidate read tests into test_metric_reader.py
Move all reading-related tests and their Metric subclass fixtures from
tests/test_metric.py into tests/test_metric_reader.py. The tests still
call Metric.read() — this is a pure file move with no behavior change.

Drops the now-empty tests/test_metric.py.

Prepares the ground for the Metric.read() removal: with everything in
one place, the follow-up PR can flip the call sites to
MetricReader.open() in a single file.
